fix/#52 : FriendReq, FriendRes @Schema name 태그 삭제, 친구요청, 친구등록 시 중복 검사 로직 추가, ErrorCode 추가

+++ zooting-api/src/main/java/com/zooting/api/application/usecase/AcceptFriendUsecase.java
@@ -33,6 +33,10 @@ public class AcceptFriendUsecase {
 
         Friend friend1 = new Friend(member1, member2);
         Friend friend2 = new Friend(member2, member1);
+        if(friendRepository.existsByFollowerAndFollowing(member1, member2) ||
+                friendRepository.existsByFollowerAndFollowing(member2, member1)) {
+            throw new BaseExceptionHandler(ErrorCode.ALREADY_EXIST_FRIEND);
+        }
         friendRepository.saveAll(List.of(friend1, friend2));
         friendRequestRepository.deleteFriendRequestByFromAndTo(member2, member1);
     }
